COMPRESS-477 building a split/spanned zip
add testcases for building a split/spanned zip, as the PR of extracting a split zip is not merged yet,
there're no other ways to test in code if my constructed zip is valid.

I will add this part of testcases as soon as the PR of extracting split zip is merged.